# web/backend/api/docs/swagger.yaml
openapi: 3.0.0
servers:
  - url: "https://api.eesast.com"
    description: Remote API Server
  - url: "http://localhost:{port}"
    description: Local API Server
    variables:
      port:
        description: 端口号
        default: "28888"
info:
  title: API Backend for EESAST
  description: |
    清华大学电子工程系学生科协后端 API。了解更多 EESAST 内容，请访问：[GitHub 主页](https://github.com/eesast/)或 [EESAST 官方网站](https://eesast.com/)。
  version: 1.0.0
  contact:
    name: API Support
    email: i@raineggplant.com
...
paths:
...
  "/static/images/:filename":
    get:
      tags:
        - static
      summary: 下载图片
      description: 下载webp格式或其本身
      operationId: getImage
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: string
                format: binary
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          description: "Not Found: File does not exist"
        500:
          description: Server error
  "/static/thuai/:filename":
    get:
      tags:
        - static
      summary: 下载thuai相关文件（目前为回放）
      description: 下载thuai相关文件（目前为回放）
      operationId: getThuaiFile
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: string
                format: binary
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          description: "Not Found: File does not exist"
        500:
          description: Server error
  "/static/{category}":
    post:
      tags:
        - static
      summary: 上传文件
      description: 需要已登录用户
      operationId: uploadFile
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        201:
          description: Successful operation
          headers:
            Location:
              description: 新增文件的 URL
              schema:
                type: string
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        500:
          description: Server error
  "/static/{category}/{filename}":
    delete:
      tags:
        - static
      summary: 删除一个文件
      description: 需要 `root`, `writer`, `editor`, `keeper`, `organizer` 或 `counselor` 角色
      operationId: deleteFile
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: Successful operation
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: File does not exist"
        500:
          description: Server error
  "/v1/articles":
    get:
      tags:
        - articles
      summary: 返回匹配查询的文章
      description: 可以通过 query 进行筛选 查看隐藏文章需要是作者、发布者、`root`或`editor`角色
      operationId: findArticles
      parameters:
        - name: title
          in: query
          description: 文章标题，采用部分匹配
          schema:
            type: string
        - name: author
          in: query
          description: 作者
          schema:
            type: string
        - name: authorId
          in: query
          description: 作者的用户id
          schema:
            type: number
        - name: createdBy
          in: query
          description: 发布文章的用户的id
          schema:
            type: number
        - name: alias
          in: query
          description: 别名，通常是文章的 URL
          schema:
            type: string
        - name: tag
          in: query
          description: 标签
          schema:
            type: string
        - name: likedBy
          in: query
          description: 点赞者 ID
          schema:
            type: number
        - name: begin
          in: query
          description: 按时间降序排列，起始文章的位置
          schema:
            type: number
        - name: end
          in: query
          description: 按时间降序排列，结束文章的位置
          schema:
            type: number
        - name: noContent
          in: query
          description: "设置是否需要获取文章内容。若设置为 `true`, 则响应中不包含 `content` 键值对"
          schema:
            type: boolean
        - name: invisible
          in: query
          description: 设置是否查看隐藏的文章
          schema:
            type: boolean
        - name: count
          in: query
          description: 设置是否查看对应authorId或createdBy文章的数量
          schema:
            type: boolean
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
        400:
          $ref: "#/components/responses/BadRequest"
        500:
          description: Server error

    post:
      tags:
        - articles
      summary: 发布一篇新文章
      description: 需要 `root` 或 `writer` 角色
      operationId: addArticle
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/addArticle"
      responses:
        201:
          description: Successful operation
          headers:
            Location:
              description: 新发布文章的 URL
              schema:
                type: string
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        409:
          description: "Conflict: Alias already exists"
        422:
          description: "UnProcessable Entity: Missing contents"
        500:
          description: Server error
  "/v1/articles/{articleId}":
    get:
      tags:
        - articles
      summary: 根据 ID 获取文章
      operationId: getArticleById
      parameters:
        - name: articleId
          in: path
          description: 文章 ID
          required: true
          schema:
            type: number
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          description: "Not Found: Article does not exist"
        500:
          description: Server error

    put:
      tags:
        - articles
      summary: 更新已存在的文章
      description: 需要 `root` 角色，或待更新文章的作者为请求者
      operationId: updateArticle
      security:
        - bearerAuth: []
      parameters:
        - name: articleId
          in: path
          description: 待更新文章的 ID
          required: true
          schema:
            type: number
      requestBody:
        $ref: "#/components/requestBodies/updateArticle"
      responses:
        204:
          description: Successful operation
          headers:
            Location:
              description: 被更新文章的 URL
              schema:
                type: string
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        403:
          description: "Forbidden: Article is under review"
        404:
          description: "Not Found: Article does not exist"
        500:
          description: Server error

    delete:
      tags:
        - articles
      summary: 删除一篇文章
      description: 需要 `root` 角色
      operationId: deleteArticle
      security:
        - bearerAuth: []
      parameters:
        - name: articleId
          in: path
          description: 要删除文章的 ID
          required: true
          schema:
            type: number
      responses:
        204:
          description: Successful operation
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: Article does not exist"
        500:
          description: Server error
  "/v1/articles/{articleId}/like":
    get:
      tags:
        - articles
      summary: 点赞
      description: 需要已登录用户
      operationId: likeArticle
      security:
        - bearerAuth: []
      parameters:
        - name: articleId
          in: path
          description: 点赞文章的 ID
          required: true
          schema:
            type: number
      responses:
        204:
          description: Successful operation
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: Article does not exist"
        500:
          description: Server error
  "/v1/articles/{articleId}/unlike":
    get:
      tags:
        - articles
      summary: 取消点赞
      description: 需要已登录用户
      operationId: unlikeArticle
      security:
        - bearerAuth: []
      parameters:
        - name: articleId
          in: path
          description: 取消点赞文章的 ID
          required: true
          schema:
            type: number
      responses:
        204:
          description: Successful operation
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: Article does not exist"
        500:
          description: Server error
  "/v1/comments":
    get:
      tags:
        - comments
      summary: 返回匹配查询的评论
      description: 可以通过 query 进行筛选
      operationId: findComments
      parameters:
        - name: replyTo
          in: query
          description: 回复的上级评论的 ID, -1 代表一级（独立）评论
          schema:
            type: number
        - name: likedBy
          in: query
          description: 点赞者 ID
          schema:
            type: number
        - name: articleId
          in: query
          schema:
            type: number
        - name: authorId
          in: query
          schema:
            type: number
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
        400:
          $ref: "#/components/responses/BadRequest"
        500:
          description: Server error

    post:
      tags:
        - comments
      summary: 发布一条新评论
      description: "需要 `root`, `writer` 或 `reader` 角色"
      operationId: addComment
      security:
        - bearerAuth: []
      requestBody:
        $ref: "#/components/requestBodies/addComment"
      responses:
        201:
          description: Successful operation
          headers:
            Location:
              description: 新发布评论的 URL
              schema:
                type: string
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        500:
          description: Server error
  "/v1/comments/{commentId}":
    get:
      tags:
        - comments
      summary: 根据 ID 获取评论
      operationId: getCommentById
      parameters:
        - name: commentId
          in: path
          description: 评论 ID
          required: true
          schema:
            type: number
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        400:
          $ref: "#/components/responses/BadRequest"
        404:
          description: "Not Found: Comment does not exist"
        500:
          description: Server error

    put:
      tags:
        - comments
      summary: 更新已存在的评论
      description: 需要 `root` 角色，或待更新评论的作者为请求者
      operationId: updateComment
      security:
        - bearerAuth: []
      parameters:
        - name: commentId
          in: path
          description: 待更新评论的 ID
          required: true
          schema:
            type: number
      requestBody:
        $ref: "#/components/requestBodies/updateComment"
      responses:
        204:
          description: Successful operation
          headers:
            Location:
              description: 被更新评论的 URL
              schema:
                type: string
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: Comment does not exist"
        500:
          description: Server error

    delete:
      tags:
        - comments
      summary: 删除一条评论
      description: 需要 `root` 角色
      operationId: deleteComment
      security:
        - bearerAuth: []
      parameters:
        - name: commentId
          in: path
          description: 待删除评论的 ID
          required: true
          schema:
            type: number
      responses:
        204:
          description: Successful operation
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: Comment does not exist"
        500:
          description: Server error
  "/v1/comments/{commentId}/like":
    get:
      tags:
        - comments
      summary: 点赞
      description: 需要已登录用户
      operationId: likeComment
      security:
        - bearerAuth: []
      parameters:
        - name: commentId
          in: path
          description: 点赞评论的 ID
          required: true
          schema:
            type: number
      responses:
        204:
          description: Successful operation
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: Comment does not exist"
        500:
          description: Server error
  "/v1/comments/{commentId}/unlike":
    get:
      tags:
        - comments
      summary: 取消点赞
      description: 需要已登录用户
      operationId: unlikeComment
      security:
        - bearerAuth: []
      parameters:
        - name: commentId
          in: path
          description: 取消点赞评论的 ID
          required: true
          schema:
            type: number
      responses:
        204:
          description: Successful operation
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/UnauthorizedError"
        404:
          description: "Not Found: Comment does not exist"
        500:
          description: Server error
...
components:
  schemas:
    Article:
      type: object
      required:
        - id
        - title
        - alias
        - authorId
        - content
      properties:
        id:
          type: number
          description: 自增值
        title:
          type: string
        alias:
          type: string
          description: 用于 SEO 与 URL
        authorId:
          type: number
        content:
          type: string
          description: 文章内容(支持 Markdown)
        abstract:
          type: string
          description: 摘要，用于预览功能
        image:
          type: string
          description: 文章标题图片
        views:
          type: number
          default: 0
        likers:
          type: array
          items:
            type: number
            description: 点赞用户的 ID
        tags:
          type: array
          items:
            type: string
        visible:
          type: boolean
          default: false
          description: 文章可见性，用于审阅功能
        createdAt:
          $ref: "#/components/schemas/Date"
        createdBy:
          type: number
        updatedAt:
          $ref: "#/components/schemas/Date"
        updatedBy:
          type: number
    Comment:
      type: object
      required:
        - id
        - authorId
        - articleId
        - content
        - replyTo
      properties:
        id:
          type: number
          description: 自增值
        authorId:
          type: number
        articleId:
          type: number
        content:
          type: string
          description: 评论内容(支持 Markdown)
        replyTo:
          type: number
          description: "回复的上级评论的 ID, -1 代表一级（独立）评论"
        likers:
          type: array
          items:
            type: number
            description: 点赞用户的 ID
        createdAt:
          $ref: "#/components/schemas/Date"
        createdBy:
          type: number
        updatedAt:
          $ref: "#/components/schemas/Date"
        updatedBy:
          type: number
...
    Date:
      type: string
      format: date-time
      description: >-
        日期对象。使用时采用 ISO
        格式字符串表示。[更多信息](https://www.iso.org/iso-8601-date-and-time-format.html)
...
  requestBodies:
    addArticle:
      content:
        application/json:
          schema:
            type: object
            required:
              - title
              - alias
              - authorId
              - content
            properties:
              title:
                type: string
              alias:
                type: string
                description: 用于 SEO 与 URL
              authorId:
                type: number
              content:
                type: string
                description: 文章内容(支持 Markdown)
              abstract:
                type: string
                description: 摘要，用于预览功能
              image:
                type: string
                description: 文章标题图片
              tags:
                type: array
                items:
                  type: string
              visible:
                type: boolean
                default: false
                description: 文章可见性，用于审阅功能
      description: 发布文章的提交格式
      required: true
    updateArticle:
      content:
        application/json:
          schema:
            type: object
            properties:
              title:
                type: string
              alias:
                type: string
                description: 用于 SEO 与 URL
              authorId:
                type: number
              content:
                type: string
                description: 文章内容(支持 Markdown)
              abstract:
                type: string
                description: 摘要，用于预览功能
              image:
                type: string
                description: 文章标题图片
              views:
                type: number
                default: 0
              likers:
                type: array
                items:
                  type: number
                  description: 点赞用户的 ID
              tags:
                type: array
                items:
                  type: string
              visible:
                type: boolean
                default: false
                description: 文章可见性，用于审阅功能
      description: 更新文章的提交格式
      required: true
    addComment:
      content:
        application/json:
          schema:
            type: object
            required:
              - authorId
              - articleId
              - content
              - replyTo
            properties:
              authorId:
                type: number
              articleId:
                type: number
              content:
                type: string
                description: 评论内容(支持 Markdown)
              replyTo:
                type: number
                description: "回复的上级评论的 ID, -1 代表一级（独立）评论"
      description: 发布评论的提交格式
      required: true
    updateComment:
      content:
        application/json:
          schema:
            type: object
            properties:
              authorId:
                type: number
              articleId:
                type: number
              content:
                type: string
                description: 评论内容(支持 Markdown)
              replyTo:
                type: number
                description: "回复的上级评论的 ID, -1 代表一级（独立）评论"
              likers:
                type: array
                items:
                  type: number
                  description: 点赞用户的 ID
      description: 更新评论的提交格式
      required: true
...
  responses:
    BadRequest:
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                description: 错误列表
                items:
                  type: object
                  properties:
                    path:
                      type: string
                      description: 错误位置
                      example: ".params.articleId"
                    message:
                      type: string
                      description: 错误信息
                      example: "should be number"
                    errorCode:
                      type: string
                      description: 错误码
                      example: "type.openapi.validation"
      description: 包含请求错误信息的 JSON
    UnauthorizedError:
      description: "Access token is missing, invalid or expired / Permission denied"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT